= core: use org.HdrHistogram.Recorder instead of our custom equivalent.
The org.HdrHistogram.Recorder class provides pretty much the same that
we needed when initially created our HdrHistogram implementation, more
specifically the posibility to take snapshots from a live Histogram
without requiring blocking the access to writes on the Histogram and
leaving it in a consistent state after finishing the snapshot operation.

This will certainly double memory consumption, as for all the Histogram
based metrics there will be two real Histograms (active and inactive),
but people should have thousands of Histograms for this to become an issue
and so far, our users have lower hundreds of Histograms in their apps.

I also removed the compact records, meaning that snapshots are now
double the size that they used to be. The access code for reading the
snapshot data is simpler, since no calculation is needed as it did
before, it's just sequential access over a Long array (it also was
sequential access before, but with some bit ops to get the real numbers).